requirements.txt
openpyxl>=3.0.9
Pillow>=10.2.0
opencv-python>=4.8.1.78
pytesseract>=0.3.8
numpy>=1.21.0
